ms_to_ticks use uint16_t instead of uint32_t

+++ src/timer_ms_to_ticks.c
@@ -3,9 +3,9 @@
 
 uint16_t timer_ms_to_ticks(uint16_t ms, timer_clock_divider_t divider, uint8_t frequency_divider_scale)
 {
-  uint32_t mcu_freq_ms = MAIN_FoscH >> frequency_divider_scale;
-  uint32_t timer_freq_ms = mcu_freq_ms / divider;
-  uint32_t ticks_count = ms * timer_freq_ms;
+  uint16_t mcu_freq_ms = MAIN_FoscH >> frequency_divider_scale;
+  uint16_t timer_freq_ms = mcu_freq_ms / divider;
+  uint16_t ticks_count = ms * timer_freq_ms;
 
-  return (ticks_count >= UINT16_MAX) ? 0 : (uint16_t) ticks_count;
+  return (ticks_count >= UINT16_MAX) ? 0 : ticks_count;
 }
